Refactor: moved text/ to style_bert_vits2/text_processing/(language)/
