+++ style_bert_vits2/text_processing/__init__.py
@@ -67,18 +67,17 @@ def clean_text(
 
     # Changed to import inside if condition to avoid unnecessary import
     if language == Languages.JP:
-        from style_bert_vits2.text_processing.japanese.g2p import g2p
-        from style_bert_vits2.text_processing.japanese.normalizer import normalize_text
+        from style_bert_vits2.text_processing.japanese import g2p, normalize_text
         norm_text = normalize_text(text)
         phones, tones, word2ph = g2p(norm_text, use_jp_extra, raise_yomi_error)
     elif language == Languages.EN:
-        from ...text import english as language_module
-        norm_text = language_module.normalize_text(text)
-        phones, tones, word2ph = language_module.g2p(norm_text)
+        from style_bert_vits2.text_processing.english import g2p, normalize_text
+        norm_text = normalize_text(text)
+        phones, tones, word2ph = g2p(norm_text)
     elif language == Languages.ZH:
-        from ...text import chinese as language_module
-        norm_text = language_module.normalize_text(text)
-        phones, tones, word2ph = language_module.g2p(norm_text)
+        from style_bert_vits2.text_processing.chinese import g2p, normalize_text
+        norm_text = normalize_text(text)
+        phones, tones, word2ph = g2p(norm_text)
     else:
         raise ValueError(f"Language {language} not supported")
