+++ style_bert_vits2/text_processing/japanese/__init__.py
@@ -1,0 +1,2 @@
+from style_bert_vits2.text_processing.japanese.g2p import g2p  # type: ignore
+from style_bert_vits2.text_processing.japanese.normalizer import normalize_text  # type: ignore
